Implement logic for user to favorite/unfavorite movies
