+++ config/routes.rb
@@ -11,6 +11,13 @@ Rails.application.routes.draw do
   get    '/login', to: 'sessions#new'
   post   '/login', to: 'sessions#create'
   delete '/logout',to: 'sessions#destroy'
+
+  resources :users do
+    member do
+      get :favorites
+    end
+  end
   
-  resources :users, :movies
+  resources :movies
+  resources :favorites, only: [:create, :destroy]
 end
